Update .readthedocs.yaml due to build component deprecation
RTD showed a message saying that we need to edit this file due to deprecation of the build:image setting in favor of build:os. I suppose the setting was used as a hidden default because it is not actually in the file. My interpretation is that we need to explicitly add a build section with the new setting.
More info:
https://blog.readthedocs.com/use-build-os-config/